pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "projectman"
version = "0.8.0"
description = "Git-native project management for Claude Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jinja2>=3.0",
    "python-frontmatter>=1.0",
]

[project.optional-dependencies]
mcp = ["mcp[cli]>=1.0"]
embeddings = ["fastembed>=0.4", "numpy>=1.24"]
web = ["fastapi>=0.115", "uvicorn[standard]>=0.34"]
all = ["projectman[mcp,embeddings,web]"]
dev = ["pytest>=7.0", "pytest-tmp-files>=0.0.2", "httpx>=0.27"]

[project.scripts]
projectman = "projectman.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/projectman"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
